requirements.txt
python-dotenv
pydantic
langchain-core
langchain-google-genai
